+++ packages/client-reconciliation/RELEASES.md
@@ -119,3 +119,13 @@ Based on:
 - [typescript v1.0.0-alpha.9] packages/client-reconciliation
 ### Releases
 - [NPM v1.0.0-alpha.9] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.9 - packages/client-reconciliation
+
+## 2024-08-13 00:16:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.363.1 (2.396.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.10] packages/client-reconciliation
+### Releases
+- [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.10 - packages/client-reconciliation